correction

+++ 2-do_deploy_web_static.py
@@ -1,30 +1,36 @@
 #!/usr/bin/python3
+""" Deploys archive!
 """
-Script that distributes archive to webservers
-"""
+from fabric.api import env, put, run
 from os.path import exists
-from fabric.operations import env, run, put, sudo
-
 env.hosts = ['54.237.112.44', '35.175.63.68']
+env.user = 'ubuntu'
 
 
 def do_deploy(archive_path):
-    """Copies archive file from local to my webservers"""
+    """ 
+    Distributes a .tgz archive from the contents of `web_static/` in AirBnB
+    clone repo to the web servers
+    Retuns:
+        (bool): `True` if all operations successful, `False` otherwise
+    """
+    if not exists(archive_path) or archive_path is None:
+        return False
 
-    if not exists(archive_path):
-        return False
-    try:
-        put(archive_path, '/tmp/')
-        file_name = archive_path.split('/')[-1].split('.')[0]
-        file_folder = '/data/web_static/releases/{file_name}'
-        run(f"mkdir -p {file_folder}")
-        run(f"tar -xzf /tmp/{file_name}.tgz -C {file_folder}")
-        run(f'rm /tmp/{file_name}.tgz')
-        run(f'mv {file_folder}/web_static/* {file_folder}')
-        run(f'rm -rf {file_folder}/web_static')
-        run('rm -rf /data/web_static/current')
-        run(f'ln -s {file_folder}/ /data/web_static/current')
-        print('New version deployed!')
-        return True
-    except Exception:
-        return False
+    file_name = archive_path.split('/')[-1]
+    dir_name = file_name.split('.')[0]
+
+    put(local_path=archive_path, remote_path='/tmp/')
+    run('mkdir -p /data/web_static/releases/{}/'.format(dir_name))
+    run('tar -xzf /tmp/{} -C /data/web_static/releases/{}/'.format(
+        file_name, dir_name))
+    run('rm /tmp/{}'.format(file_name))
+    run('mv /data/web_static/releases/{}/web_static/* '.format(dir_name) +
+        '/data/web_static/releases/{}/'.format(dir_name))
+    run('rm -rf /data/web_static/releases/{}/web_static'.format(dir_name))
+    run('rm -rf /data/web_static/current')
+    run('ln -s /data/web_static/releases/{}/ /data/web_static/current'.format(
+        dir_name))
+
+    print('New version deployed!')
+    return True
